app/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-magic
markitdown[all]
python-multipart
